refactor: watch route name change at root component
