+++ frontend/src/App.vue
@@ -24,8 +24,8 @@
 </template>
 
 <script lang="ts" setup>
-import { reactive, watchEffect, onErrorCaptured } from "vue";
-import { useRouter } from "vue-router";
+import { reactive, watchEffect, onErrorCaptured, watch } from "vue";
+import { useRoute, useRouter } from "vue-router";
 
 import { isDev } from "./utils";
 import { BBNotificationItem } from "./bbkit/types";
@@ -36,7 +36,13 @@ import HelpDrawer from "@/components/HelpDrawer.vue";
 import { NConfigProvider, NDialogProvider } from "naive-ui";
 import { themeOverrides, dateLang, generalLang } from "../naive-ui.config";
 import { t } from "./plugins/i18n";
-import { useAuthStore, useNotificationStore } from "./store";
+import {
+  useAuthStore,
+  useNotificationStore,
+  useUIStateStore,
+  useHelpStore,
+} from "./store";
+import { RouteMapList } from "@/types";
 // Show at most 3 notifications to prevent excessive notification when shit hits the fan.
 const MAX_NOTIFICATION_DISPLAY_COUNT = 3;
 
@@ -49,15 +55,20 @@ const CRITICAL_NOTIFICATION_DURATION = 10000;
 interface LocalState {
   notificationList: BBNotificationItem[];
   prevLoggedIn: boolean;
+  helpTimer: number | null;
+  RouteMapList: RouteMapList | null;
 }
 
 const authStore = useAuthStore();
 const notificationStore = useNotificationStore();
+const route = useRoute();
 const router = useRouter();
 
 const state = reactive<LocalState>({
   notificationList: [],
   prevLoggedIn: authStore.isLoggedIn(),
+  helpTimer: null,
+  RouteMapList: null,
 });
 
 setInterval(() => {
@@ -109,6 +120,40 @@ const watchNotification = () => {
     }
   }
 };
+
+// watch route change for help
+watch(
+  () => route.name,
+  async (routeName) => {
+    const uiStateStore = useUIStateStore();
+    const helpStore = useHelpStore();
+
+    if (!state.RouteMapList) {
+      const res = await fetch("/help/routeMapList.json");
+      state.RouteMapList = await res.json();
+    }
+
+    // Clear timer after every route change.
+    if (state.helpTimer) {
+      clearTimeout(state.helpTimer);
+      state.helpTimer = null;
+    }
+
+    const helpName = state.RouteMapList?.find(
+      (pair) => pair.routeName === routeName
+    )?.helpName;
+
+    if (helpName && !uiStateStore.getIntroStateByKey(`guide.${helpName}`)) {
+      state.helpTimer = window.setTimeout(() => {
+        helpStore.showHelp(helpName, true);
+        uiStateStore.saveIntroStateByKey({
+          key: "environment.visit",
+          newState: true,
+        });
+      }, 1000);
+    }
+  }
+);
 
 watchEffect(watchNotification);
