+++ frontend/src/router/index.ts
@@ -30,13 +30,7 @@ import {
   Sheet,
   UNKNOWN_ID,
 } from "../types";
-import {
-  idFromSlug,
-  isDBAOrOwner,
-  isOwner,
-  isProjectOwner,
-  handleRouteChangedForHelp,
-} from "../utils";
+import { idFromSlug, isDBAOrOwner, isOwner, isProjectOwner } from "../utils";
 // import PasswordReset from "../views/auth/PasswordReset.vue";
 import Signin from "../views/auth/Signin.vue";
 import Signup from "../views/auth/Signup.vue";
@@ -1344,7 +1338,7 @@ router.beforeEach((to, from, next) => {
   });
 });
 
-router.afterEach((to, from) => {
+router.afterEach((to /*, from*/) => {
   // Needs to use nextTick otherwise title will still be the one from the previous route.
   nextTick(() => {
     if (to.meta.title) {
@@ -1353,7 +1347,4 @@ router.afterEach((to, from) => {
       document.title = "Bytebase";
     }
   });
-  if (to.name !== from.name) {
-    handleRouteChangedForHelp(to);
-  }
 });
